# Conflicts:
#	src/main/java/de/dafuqs/spectrum/api/energy/color/InkColor.java
#	src/main/java/de/dafuqs/spectrum/blocks/conditional/colored_tree/ColoredFenceGateBlock.java
#	src/main/java/de/dafuqs/spectrum/blocks/conditional/colored_tree/ColoredStrippedLogBlock.java
#	src/main/java/de/dafuqs/spectrum/blocks/conditional/colored_tree/ColoredStrippedWoodBlock.java

+++ src/main/java/de/dafuqs/spectrum/helpers/enchantments/ImprovedCriticalHelper.java
@@ -9,7 +9,7 @@ import net.minecraft.world.item.*;
 public class ImprovedCriticalHelper {
 	
 	public static float getAdditionalCritDamageMultiplier(int improvedCriticalLevel) {
-		return SpectrumConfig.CONFIG.ImprovedCriticalExtraDamageMultiplierPerLevel.get() * improvedCriticalLevel;
+		return SpectrumConfig.CONFIG.ImprovedCriticalExtraDamageMultiplierPerLevel.get().floatValue() * improvedCriticalLevel;
 	}
 	
 	public static float getAdditionalCritDamageMultiplier(HolderLookup.Provider lookup, ItemStack stack) {
